Added word wrapping to line parser
