+++ ui/window/help/help.go
@@ -4,6 +4,7 @@ package help
 // Options to open to a specific help page.
 
 import (
+	"github.com/yamamushi/EscapingEden/ui/config"
 	"github.com/yamamushi/EscapingEden/ui/types"
 	"github.com/yamamushi/EscapingEden/ui/util"
 	"github.com/yamamushi/EscapingEden/ui/window"
@@ -29,7 +30,7 @@ type HelpWindow struct {
 
 func NewHelpWindow(x, y, w, h, consoleWidth, consoleHeight int, page types.HelpPage, input, output chan string) *HelpWindow {
 	hw := &HelpWindow{}
-	hw.ID = window.HELPBOX
+	hw.ID = config.WindowHelpBox
 	// if x or y are less than 1 set them to 1
 	if x < 1 {
 		x = 1
